Curve.h ESP8266

+++ Curve.h
@@ -11,11 +11,11 @@
 #include <pgmspace.h>
 
 class Curve {
- static const uint8_t etable[] PROGMEM;
+ static const uint16_t etable[] PROGMEM;
 public:
- static uint8_t exponential(uint8_t);
- static uint8_t linear(uint8_t);
- static uint8_t reverse(uint8_t);
+ static uint16_t exponential(uint8_t);
+ static uint16_t linear(uint8_t);
+ static uint16_t reverse(uint8_t);
 };
 
 #endif	/* CURVE_H */
